requirements.txt
discord
python-dotenv==0.19.0